requirements.txt
# Derin öğrenme
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

# Görüntü işleme
Pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.1

# Makine öğrenmesi metrikleri
scikit-learn>=1.3.0

# Grad-CAM
grad-cam>=1.4.8

# Veri seti
huggingface_hub>=0.20.0
datasets>=2.16.0

# Görselleştirme
matplotlib>=3.7.0
seaborn>=0.12.0

# Yardımcı
numpy>=1.24.0
tqdm>=4.66.0